services/training/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-training"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "mlflow>=2.11.0,<2.18.0",
    "psycopg2-binary>=2.9.9",
    "structlog>=24.1.0",
    "minio>=7.2.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.17.0",
    "onnxscript>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
